% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_configfile.R
\name{gs_configfile}
\alias{gs_configfile}
\title{Creates config file
Creates a template for a config file to be used in genomic selection with multiple models and multiple phenotypes.}
\usage{
gs_configfile(type, configDir = ".")
}
\arguments{
\item{type}{"multi" or "single" for multiple phenotype or single phenotype configuration file.}

\item{configDir}{(Optional) Directory where filenames are locates. Default dir is "."}
}
\value{
None
}
\description{
Creates config file
Creates a template for a config file to be used in genomic selection with multiple models and multiple phenotypes.
}
