% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_datasets.R
\name{gs_datasets}
\alias{gs_datasets}
\title{Write datasets to files}
\usage{
gs_datasets(type)
}
\arguments{
\item{type}{String defining the type of dataset to generate: "small", "full", "sources, or "single".
"small" and "full" create both a configuration file (YAML format) and training/target datasets for genomic selection with multiple phentotypes. The former with a genotype of 1000 markers and a phenotype of 6 traits, and the latest with a genotype of 4640 markers and 24 traits.
"sources" copies original genotype and phenotype datasets used to generate the "small" and "full" datasets.
And "single" creates both a configuration file (YAML format) and training/target datasets for genomic selection with single phenotype.}
}
\value{
None. Files are written into directories with the same name of the type arguments (i.e, "small", "full", "sources", and "single")
}
\description{
Function that writes into CSV files different types of datasets used for genomica selection (training/target/sources phenotypes and genotypes)
}
