% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_heritability.R
\name{gs_heritability}
\alias{gs_heritability}
\title{Calculates and plots heritabilities for HCL penotypes.}
\usage{
gs_heritability(genoFile, phenosFile, outputDir)
}
\arguments{
\item{genoFile}{Filename of genotype matrix.}

\item{phenosFile}{Filename of phenotypes table.}

\item{outputDir}{Dirname where outputs will be saved.}
}
\value{
None. Generate two files: a table (CSV) with heritabilities for each trait
and a summary plot (PDF) comparing heritabilities from all traits.
}
\description{
Calculates and plots heritabilities for HCL penotypes.
}
