% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_multi.R
\name{gs_multi}
\alias{gs_multi}
\title{Genomic selection for multiple phenotypes and multiple models}
\usage{
gs_multi(configFile, outputDir)
}
\arguments{
\item{configFile}{Filename of the parameter file (YAML Format).  It includes:
trainingGenoFile: Filename of training genotype data,
trainingPhenoFile: Filename of training phenotype data,
trainingGenoFile: Filename of target genotype table,
targetPhenoFile: Filename of target phenotype table (opt.),
markersFile: Filename of selected markers to test from genotype (opt.),
nMarkers: Number of selected markers to test,
methods: Methods or models to test in GS,
nTimes: Number of independent replicates for the cross-validation,
nFolds: Number of folds for the cross-validation,
nCores: Number of cores to use.}

\item{outputDir}{The directory where the results will be saved.}
}
\value{
None
}
\description{
Runs in parallel genomic selection for multiple phenotypes and multiple models.
}
