% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_plot_distributions.R
\name{gs_plot_distributions}
\alias{gs_plot_distributions}
\title{Create a plot for color trait densitydistributioons.}
\usage{
gs_plot_distributions(HCLFilename)
}
\arguments{
\item{filename}{of table with values of HCL trait.}
}
\description{
Create a plot for color trait densitydistributioons.
}
