% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gs_plots_varying_markers}
\alias{gs_plots_varying_markers}
\title{Create summaries when the number of markers varies.}
\usage{
gs_plots_varying_markers(inputDir)
}
\arguments{
\item{inputDir}{Dir name with results from different GS runs.}
}
\value{
None. Writes table (.csv) and plots (.pdfs) to files.
}
\description{
Create one summary table and two plots: one per base trait and one per hcl trait.
}
