% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_plotsHCL.R
\name{gs_plotsHCL}
\alias{gs_plotsHCL}
\title{Create a comparison plot for potato color phenotypes.}
\usage{
gs_plotsHCL(HCLFilename, outputDir, columnName, TITLE)
}
\arguments{
\item{HCLTable}{Table with information from cross validation: GEBVs for each fold and each repetition.}
}
\value{
None. Plot is saved to a PDF file
}
\description{
Function that creates boxplots comparing HCL color components phenotypes (e.g. Heritability or Prediction_ability).
}
