% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_single.R
\name{gs_single}
\alias{gs_single}
\title{Genomic selection for single phenotype and multiple models}
\usage{
gs_single(paramsFile, outputDir = "")
}
\arguments{
\item{paramsFile}{Filename of the parameter file (YAML Format).}

\item{outputDir}{(Optional) Directory name where results will be saved. If empty, phenotype name is used.}
}
\value{
TRUE if no errors or FALSE otherwise.
}
\description{
Function that performs genomic selection by evaluating in parallel multiple models on a single phenotype.
Models are evaluated through cross-validation, then the best model is selected,
and GEBVs are calculated using this model.
Results are generated in both tables and graphics, and they include:
GEBVs for target and training datasets, graphic comparison of predicted ability for each model
(correlation between actual and predicted values), and comparison of execution times between models.
}
