% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_split.R
\name{gs_split}
\alias{gs_split}
\title{Split genotype/phenotype datasets into training and target datasets}
\usage{
gs_split(genoFile, phenosFile, outputDir, nGenos = NA, nPhenos = NA)
}
\arguments{
\item{genoFile}{Filename of genotype data matrix (rows for markers, columns for samples)}

\item{phenosFile}{Filename of fenotype data table (rows for samples, columns for phenotypes)}

\item{outputDir}{The directory where the results will be saved.}

\item{nGenos}{(Optional) Number of markers to use from genotype data matrix. If empty all markers are used.}

\item{nPhenos}{(Optional) Number of phenotype to use from phenotype data table. If empty all phenotypes are used.}
}
\value{
None, results are saved to the output directory
}
\description{
Function that splits a complete genotype/phenotypes data into training/target datasets.
}
