% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_plots.R, R/gs_summaryPlot.R
\name{gs_summaryPlot}
\alias{gs_summaryPlot}
\title{Summary plot for cross validation.}
\usage{
gs_summaryPlot(predictionsKFoldsFile)

gs_summaryPlot(predictionsKFoldsFile)
}
\arguments{
\item{predictionsKFoldsFile}{Filename with data from cross validation: GEBVs for each fold and each repetition.}

\item{predictionsKFoldsTable}{Table with information from cross validation: GEBVs for each fold and each repetition.}
}
\value{
None

None. Writes two files into working dir: data table (.csv) and boxplot graphics (.pdf).
}
\description{
Creates a summary ploy by comparing predictive abilities for each phenotype and each model

Creates a summary ploy by comparing predictive abilities for each phenotype and each model
}
